+++ platform/challenge/service.go
@@ -2,6 +2,8 @@ package challenge
 
 import (
 	"context"
+	"errors"
+
 	challengestore "github.com/micromdm/scep/challenge/bolt"
 )
 
@@ -14,6 +16,9 @@ type ChallengeService struct {
 }
 
 func (c *ChallengeService) SCEPChallenge(ctx context.Context) (string, error) {
+	if c.scepChallengeStore == nil {
+		return "", errors.New("SCEP challenge store missing")
+	}
 	return c.scepChallengeStore.SCEPChallenge()
 }